mainframe/requirements.txt
python-dotenv
grpcio
protobuf
langchain-community
langchain
google-ai-generativelanguage
google-generativeai
uvicorn
fastapi